% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LRT_STAT}
\alias{LRT_STAT}
\title{LRT_STAT, required by mlrt_test}
\usage{
LRT_STAT(n, x, s, seed)
}
\arguments{
\item{n}{... as above}

\item{x}{...}

\item{s}{...}

\item{seed}{optional, an integer that is the starting point used in the generation of a sequence of random numbers. Include a seed if you want reproducible output.}
}
\value{
xx
}
\description{
LRT_STAT, required by mlrt_test
}
