% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mlrt_test}
\alias{mlrt_test}
\title{modified signed-likelihood ratio test (SLRT) for equality of CVs}
\usage{
mlrt_test(nr = 1000, x, y)
}
\arguments{
\item{nr}{numeric vector length one, number of simulation runs, default is 1e3}

\item{x}{a numeric vector containing individual measurement values}

\item{y}{a vector of any type containing a grouping variable}
}
\value{
a list with the test statistic and p-value
}
\description{
modified signed-likelihood ratio test (SLRT) for equality of CVs
}
\examples{

 x <- rnorm(100)
 y <- unlist(lapply(letters[1:5], function(i) rep(i, 20)))

 mlrt_test(nr = 1e3, x, y)


}

