% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{mlrt_test2}
\alias{mlrt_test2}
\title{modified signed-likelihood ratio test (SLRT) for equality of CVs}
\usage{
mlrt_test2(nr, n, x, s)
}
\arguments{
\item{nr}{numeric vector lenght one, number of simulation runs}

\item{n}{a numeric vector, the number of observations in each group}

\item{x}{a numeric vector, the mean of each group}

\item{s}{a numeric vector, the standard deviation of each group}
}
\value{
a list with the test statistic and p-value
}
\description{
modified signed-likelihood ratio test (SLRT) for equality of CVs
}
\examples{

# Summary stats from Feltz and Miller 1996
#  \\url{http://onlinelibrary.wiley.com.sci-hub.cc/doi/10.1002/(SICI)1097-0258(19960330)15:6\%3C647::AID-SIM184\%3E3.0.CO;2-P/epd} (originally \\url{http://www.tandfonline.com.sci-hub.cc/doi/abs/10.1080/15321819308019845?journalCode=ljii19}

miller <- data.frame(test = c('ELISA', 'WEHI', '`Viral inhibition`'),
                    Mean = c(6.8, 8.5, 6.0),
                    CV =   c(0.090, 0.462, 0.340),
                    N =    c(5, 5, 5))
# compute SD from mean and cv
miller$SD <- with(miller, CV * Mean)

 mlrt_test2(nr = 1e3, n = miller$N, s = miller$SD, x = miller$Mean)


}

