% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{LRT_STAT}
\alias{LRT_STAT}
\title{LRT_STAT, required by mlrt_test}
\usage{
LRT_STAT(n, x, s)
}
\arguments{
\item{n}{... as above}

\item{x}{...}

\item{s}{...}
}
\value{
xx
}
\description{
LRT_STAT, required by mlrt_test
}
